requirements.txt

viam-sdk==0.69.0

typing-extensions
google-auth
google-auth-oauthlib
google-api-python-client
pyinstaller
